requirements.txt
# Núcleo RAG e LlamaIndex
llama-index==0.11.20
llama-index-llms-ollama==0.3.4
llama-index-embeddings-huggingface==0.3.1
llama-index-vector-stores-faiss==0.2.2

# Utilitários
faiss-cpu==1.8.0
torch==2.3.0
sentence-transformers==2.6.1
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2

# Sistema e performance
numpy==1.26.4
tqdm==4.66.5

# (Opcional: aceleração GPU)
# faiss-gpu==1.8.0
